Move vectordb files to workspace
