+++ packages/server/src/sdk/index.ts
@@ -23,7 +23,7 @@ import * as workspaceApps from "./workspace/workspaceApps"
 import { default as workspaces } from "./workspace/workspaces"
 import * as aiConfigs from "./aiConfig"
 import * as ragConfigs from "./ragConfig"
-import * as vectorDbs from "./vectorDb"
+import * as vectorDbs from "./workspace/ai/vectorDb"
 
 const sdk = {
   backups,
