+++ packages/server/src/sdk/workspace/ai/rag/index.ts
@@ -3,7 +3,7 @@ import * as crypto from "crypto"
 import { PDFParse } from "pdf-parse"
 import { parse as parseYaml } from "yaml"
 import sdk from "../../.."
-import { createVectorDb, type ChunkInput } from "../vectorDb"
+import { createVectorDb, type ChunkInput } from "../vectorDb/utils"
 import { HTTPError } from "@budibase/backend-core"
 import { ai } from "@budibase/pro"
 import { embedMany } from "ai"
