+++ packages/server/src/sdk/workspace/ai/vectorDb/index.ts
@@ -1,8 +1,2 @@
-import { PgVectorDb } from "./pgVectorDb"
-import type { VectorDb, VectorDbConfig } from "./types"
-
-export const createVectorDb = (config: VectorDbConfig): VectorDb => {
-  return new PgVectorDb(config)
-}
-
-export * from "./types"
+export * from "./crud"
+export * from "./utils"
