+++ packages/server/src/sdk/workspace/ai/vectorDb/utils.ts
@@ -1,0 +1,8 @@
+import { PgVectorDb } from "./pgVectorDb"
+import type { VectorDb, VectorDbConfig } from "./types"
+
+export const createVectorDb = (config: VectorDbConfig): VectorDb => {
+  return new PgVectorDb(config)
+}
+
+export * from "./types"
